+++ src/main/kotlin/blackjack/domain/player/Name.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 class Name(val value: String) {
     init {
