+++ src/main/kotlin/blackjack/domain/state/State.kt
@@ -1,7 +1,7 @@
 package blackjack.domain.state
 
-import blackjack.domain.Hands
 import blackjack.domain.card.Card
+import blackjack.domain.player.Hands
 
 interface State {
     val hands: Hands
